analytics-service/requirements.txt
django
djangorestframework
django-cors-headers
pika
psycopg2-binary
gunicorn
python-consul